Correct comments, add signals to sensitivity list
The process that multiplexes the serial data to the DAC was missing
dsdata1_left and dsdata1_right from the sensitivity list. Combinational
processes should have all signals being read inside the process in the
sensitivity list. For some reason, the code works when synthesized.
It does not work when simulated, which is expected because the dsdata
signals were not in the sensitivity list.