Change AD1939 to use 24 bits instead of 32 (#15)
* change streaming data type to 24 bits instead of 32

* Formatting and bug fix

An assignment had the wrong number of bits on the right side, which ghdl didn't report

* Remove error signal and change max channel to 1

The codec is only set up for stereo (channel 0 and channel 1);
setting channel = 3 when valid is not asserted made debugging with
a logic analyzer more difficult

* Correct comments, add signals to sensitivity list

The process that multiplexes the serial data to the DAC was missing
dsdata1_left and dsdata1_right from the sensitivity list. Combinational
processes should have all signals being read inside the process in the
sensitivity list. For some reason, the code works when synthesized.
It does not work when simulated, which is expected because the dsdata
signals were not in the sensitivity list.

* update audio research to 24 bits

* delete old files and remove v1 from names

* Updated AD1939 hw.tcl components to no longer reference v1 files

Co-authored-by: Dylan Wickham <wickham@flatearthinc.com>